# SpringbootRestApi_swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: employee-controller
    description: Employee Controller
paths:
  /api/employees:
    get:
      tags:
        - employee-controller
      summary: getEmployees
      operationId: getEmployeesUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Employee'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - employee-controller
      summary: saveEmployee
      operationId: saveEmployeeUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          required: false
          type: integer
          format: int64
        - name: name
          in: query
          required: false
          type: string
        - name: salary
          in: query
          required: false
          type: integer
          format: int32
        - name: department
          in: query
          required: false
          type: string
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/employees/{employeeId}':
    get:
      tags:
        - employee-controller
      summary: getEmployee
      operationId: getEmployeeUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: path
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Employee'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - employee-controller
      summary: updateEmployee
      operationId: updateEmployeeUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: employee
          description: employee
          required: true
          schema:
            $ref: '#/definitions/Employee'
        - name: employeeId
          in: path
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - employee-controller
      summary: deleteEmployee
      operationId: deleteEmployeeUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: path
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: string
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

definitions:
  Employee:
    type: object
    properties:
      department:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
      salary:
        type: integer
        format: int32